+++ ivy/container/extensions/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+from .elementwise import ContainerWithElementWiseExtensions
+from .layers import ContainerWithLayersExtensions
+from .manipulation import ContainerWithManipulationExtensions
+from .statistical import ContainerWithStatisticalExtensions
